move file row object into model folder
