+++ backend/src/main/java/gov/cdc/usds/simplereport/api/model/fileRow/FileRow.java
@@ -1,4 +1,4 @@
-package gov.cdc.usds.simplereport.validators;
+package gov.cdc.usds.simplereport.api.model.fileRow;
 
 import gov.cdc.usds.simplereport.service.model.reportstream.FeedbackMessage;
 import java.util.List;
