+++ backend/src/main/java/gov/cdc/usds/simplereport/config/FileValidatorConfiguration.java
@@ -1,8 +1,8 @@
 package gov.cdc.usds.simplereport.config;
 
+import gov.cdc.usds.simplereport.api.model.fileRow.PatientUploadRow;
+import gov.cdc.usds.simplereport.api.model.fileRow.TestResultRow;
 import gov.cdc.usds.simplereport.validators.FileValidator;
-import gov.cdc.usds.simplereport.validators.PatientUploadRow;
-import gov.cdc.usds.simplereport.validators.TestResultRow;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
